+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 * Fixed in-memory filesystem not handling out-of-order filesystem change events
+* Fixed long-polling error caused by a promise mixup ([#110](https://github.com/LPGhatguy/rojo/issues/110))
 
 ## [0.5.0 Alpha 1](https://github.com/LPGhatguy/rojo/releases/tag/v0.5.0-alpha.1) (January 14, 2019)
 * Changed plugin UI to be way prettier